Former-commit-id: 63b935ce5caf74142d52e2fa0ba769c82e73a22f [formerly a96df6d6ad1bfe7b5392f973fcf64173e5ea36ad] [formerly 0c11f99805b786370234f05cbd3da6c3efd69337] [formerly 0c11f99805b786370234f05cbd3da6c3efd69337 [formerly 796380dd53e8a3a232f3e9e4f6b1b4b599f6597f]] [formerly adfeb1306751695abe2eaaf23d62ed4f3baa2430 [formerly 0c11f99805b786370234f05cbd3da6c3efd69337 [formerly 796380dd53e8a3a232f3e9e4f6b1b4b599f6597f] [formerly adfeb1306751695abe2eaaf23d62ed4f3baa2430 [formerly 56dc338129ed46612df4db1926e0b7e91551b60d]]]]
Former-commit-id: adfeb1306751695abe2eaaf23d62ed4f3baa2430
Former-commit-id: 7378b42d4c3658f188a8dd75702d6d9be0ad0f52 [formerly f6e448879003172c3aafecc16e87e7783e176970] [formerly 60f9feed3efc51b1179e84a664cd1775442ffe07 [formerly 12c854e7cc66f4fd55c067ee414a3797e0cac5d2]]
Former-commit-id: 60f9feed3efc51b1179e84a664cd1775442ffe07
Former-commit-id: 0371accc473390fe9993728c95b2fbbffe0da7c7

+++ ncep/gov.noaa.nws.ncep.edex.plugin.pirep/src/gov/noaa/nws/ncep/edex/plugin/pirep/PirepDecoder.java
@@ -50,6 +50,7 @@ import com.raytheon.uf.edex.decodertools.time.TimeTools;
  * Sep 26, 2011 286        qzhou       Changed reportType from int to string
  * Aug 30, 2013 2298       rjpeter     Make getPluginName abstract
  * Sep 05, 2013 2316       bsteffen    Unify pirep and ncpirep.
+ * Jul 23, 2014 3410       bclement    location changed to floats
  * </pre>
  * 
  * @author F. J. Yen
@@ -164,8 +165,8 @@ public class PirepDecoder extends AbstractDecoder {
                     record.setObsText(DecoderTools.normalizeObs(
                             parser.getReportData(), wmoHeader.getWmoHeader()));
 
-                    location.setLatitude(p.getLatitude());
-                    location.setLongitude(p.getLongitude());
+                    location.setLatitude((float) p.getLatitude());
+                    location.setLongitude((float) p.getLongitude());
                     location.setLocation(p.getLatitude(), p.getLongitude());
                     location.setStationId(parser.getReportingStationId());
